--- vnpy/app/csv_loader/base.py
@@ -1,2 +1,0 @@
-APP_NAME = "CsvLoader"
-STOPORDER_PREFIX = "STOP"
