+++ vnpy/app/csv_loader/ui/__init__.py
@@ -1,1 +1,1 @@
-from .widget import CsvLoaderWidget
+from .csv_loader_widget import CsvLoaderWidget
