+++ vnpy/app/csv_loader/ui/csv_loader_widget.py
@@ -1,17 +1,27 @@
+"""
+Author: nanoric
+
+UI is designed by Qt Designer. source at csv_loader.ui
+compile:
+```
+pyuic5 csv_loader.ui -o uic/uic_csv_loader.py
+```
+"""
 from PyQt5.QtWidgets import QFileDialog
 from vnpy.event import EventEngine
 from vnpy.trader.constant import Interval, Exchange
 from vnpy.trader.engine import MainEngine
 from vnpy.trader.ui import QtWidgets
 
-from .uic.uic_CsvLoader import Ui_CsvLoader
-from ..CsvLoader import CsvLoader
+from .uic.uic_csv_loader import Ui_CsvLoader
+from ..csv_loader import CsvLoader
 
 
 class CsvLoaderWidget(QtWidgets.QWidget):
     """"""
 
     def __init__(self, main_engine: MainEngine, event_engine: EventEngine):
+        """"""
         super().__init__()
         self.loader = CsvLoader(main_engine, event_engine)
         self.ui = Ui_CsvLoader()
@@ -19,6 +29,7 @@ class CsvLoaderWidget(QtWidgets.QWidget):
         self.init_ui()
 
     def init_ui(self):
+        """"""
         self.ui.setupUi(self)
 
         for i in Interval:
@@ -28,12 +39,14 @@ class CsvLoaderWidget(QtWidgets.QWidget):
             self.ui.exchange_combo.addItem(str(i), i)
 
     def on_choose_button_pressed(self):
+        """"""
         result: str = QFileDialog.getOpenFileName(self)
         filename = result[0]
         if filename:
             self.ui.file_edit.setText(filename)
 
     def on_load_button_pressed(self):
+        """"""
         self.loader.file_path = self.ui.file_edit.text()
         self.loader.symbol = self.ui.symbol_edit.text()
         self.loader.exchange = self.ui.exchange_combo.currentData()
