--- vnpy/app/csv_loader/ui/ui.py
@@ -1,120 +1,0 @@
-from PyQt5 import QtCore, QtWidgets
-from vnpy.trader.constant import Exchange, Interval
-
-
-class CsvLoaderUI(object):
-
-    def __init__(self, instance: QtWidgets.QDialog):
-        instance.setObjectName("CsvLoader")
-
-        self.vertical_layout = QtWidgets.QVBoxLayout(instance)
-        self.vertical_layout.setObjectName("verticalLayout")
-        self.horizontal_layout = QtWidgets.QHBoxLayout()
-        self.horizontal_layout.setObjectName("horizontalLayout")
-        self.choose_button = QtWidgets.QPushButton(instance)
-        self.choose_button.setObjectName("choose_button")
-        self.horizontal_layout.addWidget(self.choose_button)
-        self.file_edit = QtWidgets.QLineEdit(instance)
-        self.file_edit.setObjectName("file_edit")
-        self.horizontal_layout.addWidget(self.file_edit)
-        self.vertical_layout.addLayout(self.horizontal_layout)
-        self.form_layout = QtWidgets.QFormLayout()
-        self.form_layout.setObjectName("formLayout")
-
-        self.label_8 = QtWidgets.QLabel(instance)
-        self.label_8.setObjectName("label_8")
-        self.form_layout.setWidget(0, QtWidgets.QFormLayout.FieldRole, self.label_8)
-        self.label_9 = QtWidgets.QLabel(instance)
-        self.label_9.setObjectName("label_9")
-        self.form_layout.setWidget(1, QtWidgets.QFormLayout.LabelRole, self.label_9)
-        self.symbol_edit = QtWidgets.QLineEdit(instance)
-        self.symbol_edit.setObjectName("symbol_edit")
-        self.form_layout.setWidget(1, QtWidgets.QFormLayout.FieldRole, self.symbol_edit)
-        self.label_12 = QtWidgets.QLabel(instance)
-        self.label_12.setObjectName("label_12")
-        self.form_layout.setWidget(2, QtWidgets.QFormLayout.LabelRole, self.label_12)
-        self.exchange_combo = QtWidgets.QComboBox(instance)
-        self.exchange_combo.setObjectName("exchange_combo")
-        self.form_layout.setWidget(2, QtWidgets.QFormLayout.FieldRole, self.exchange_combo)
-        self.label_11 = QtWidgets.QLabel(instance)
-        self.label_11.setObjectName("label_11")
-        self.form_layout.setWidget(3, QtWidgets.QFormLayout.LabelRole, self.label_11)
-        self.interval_combo = QtWidgets.QComboBox(instance)
-        self.interval_combo.setObjectName("interval_combo")
-        self.form_layout.setWidget(3, QtWidgets.QFormLayout.FieldRole, self.interval_combo)
-        self.label = QtWidgets.QLabel(instance)
-        self.label.setObjectName("label")
-        self.form_layout.setWidget(4, QtWidgets.QFormLayout.FieldRole, self.label)
-        self.label_2 = QtWidgets.QLabel(instance)
-        self.label_2.setObjectName("label_2")
-        self.form_layout.setWidget(5, QtWidgets.QFormLayout.LabelRole, self.label_2)
-        self.datetime_edit = QtWidgets.QLineEdit(instance)
-        self.datetime_edit.setObjectName("datetime_edit")
-        self.form_layout.setWidget(5, QtWidgets.QFormLayout.FieldRole, self.datetime_edit)
-        self.label_3 = QtWidgets.QLabel(instance)
-        self.label_3.setObjectName("label_3")
-        self.form_layout.setWidget(6, QtWidgets.QFormLayout.LabelRole, self.label_3)
-        self.open_edit = QtWidgets.QLineEdit(instance)
-        self.open_edit.setObjectName("open_edit")
-        self.form_layout.setWidget(6, QtWidgets.QFormLayout.FieldRole, self.open_edit)
-        self.label_4 = QtWidgets.QLabel(instance)
-        self.label_4.setObjectName("label_4")
-        self.form_layout.setWidget(7, QtWidgets.QFormLayout.LabelRole, self.label_4)
-        self.high_edit = QtWidgets.QLineEdit(instance)
-        self.high_edit.setObjectName("high_edit")
-        self.form_layout.setWidget(7, QtWidgets.QFormLayout.FieldRole, self.high_edit)
-        self.label_5 = QtWidgets.QLabel(instance)
-        self.label_5.setObjectName("label_5")
-        self.form_layout.setWidget(8, QtWidgets.QFormLayout.LabelRole, self.label_5)
-        self.low_edit = QtWidgets.QLineEdit(instance)
-        self.low_edit.setObjectName("low_edit")
-        self.form_layout.setWidget(8, QtWidgets.QFormLayout.FieldRole, self.low_edit)
-        self.label_6 = QtWidgets.QLabel(instance)
-        self.label_6.setObjectName("label_6")
-        self.form_layout.setWidget(9, QtWidgets.QFormLayout.LabelRole, self.label_6)
-        self.close_edit = QtWidgets.QLineEdit(instance)
-        self.close_edit.setObjectName("close_edit")
-        self.form_layout.setWidget(9, QtWidgets.QFormLayout.FieldRole, self.close_edit)
-        self.label_7 = QtWidgets.QLabel(instance)
-        self.label_7.setObjectName("label_7")
-        self.form_layout.setWidget(10, QtWidgets.QFormLayout.LabelRole, self.label_7)
-        self.volume_edit = QtWidgets.QLineEdit(instance)
-        self.volume_edit.setObjectName("volume_edit")
-        self.form_layout.setWidget(10, QtWidgets.QFormLayout.FieldRole, self.volume_edit)
-
-        self.vertical_layout.addLayout(self.form_layout)
-        self.load_button = QtWidgets.QPushButton(instance)
-        self.load_button.setObjectName("load_button")
-        self.vertical_layout.addWidget(self.load_button)
-
-        for i in Interval:
-            self.interval_combo.addItem(str(i), i)
-
-        for i in Exchange:
-            self.exchange_combo.addItem(str(i), i)
-
-        self.retranslate_ui(instance)
-        QtCore.QMetaObject.connectSlotsByName(instance)
-
-    def retranslate_ui(self, instance):
-        _translate = QtCore.QCoreApplication.translate
-        instance.setWindowTitle(_translate("CsvLoader", "Csv载入器"))
-        self.choose_button.setText(_translate("CsvLoader", "选择文件"))
-        self.label_8.setText(_translate("CsvLoader", "合约信息"))
-        self.label_9.setText(_translate("CsvLoader", "Symbol"))
-        self.label_11.setText(_translate("CsvLoader", "Interval"))
-        self.label_12.setText(_translate("CsvLoader", "Exchange"))
-        self.label.setText(_translate("CsvLoader", "表头属性"))
-        self.label_2.setText(_translate("CsvLoader", "Datetime"))
-        self.datetime_edit.setText(_translate("CsvLoader", "DateTime"))
-        self.label_3.setText(_translate("CsvLoader", "Open"))
-        self.open_edit.setText(_translate("CsvLoader", "Open"))
-        self.label_4.setText(_translate("CsvLoader", "High"))
-        self.high_edit.setText(_translate("CsvLoader", "High"))
-        self.label_5.setText(_translate("CsvLoader", "Low"))
-        self.low_edit.setText(_translate("CsvLoader", "Low"))
-        self.label_6.setText(_translate("CsvLoader", "Close"))
-        self.close_edit.setText(_translate("CsvLoader", "Close"))
-        self.label_7.setText(_translate("CsvLoader", "Volume"))
-        self.volume_edit.setText(_translate("CsvLoader", "Volume"))
-        self.load_button.setText(_translate("CsvLoader", "载入"))
